+++ example/inc/adc.h
@@ -1,0 +1,14 @@
+/*
+ * adc.h
+ *
+ *  Created on: 18 авг. 2017 г.
+ *      Author: kvaoar
+ */
+
+#ifndef INC_ADC_H_
+#define INC_ADC_H_
+
+#include "stm8s.h"
+uint32_t adc_get();
+
+#endif /* INC_ADC_H_ */
