+++ example/src/adc.c
@@ -1,0 +1,12 @@
+/*
+ * adc.c
+ *
+ *  Created on: 18 авг. 2017 г.
+ *      Author: kvaoar
+ */
+#include "adc.h"
+
+uint32_t adc_get(){
+
+return 0;
+}
